Moved the UI image class to a separate file
