+++ src/common/CMakeLists.txt
@@ -72,6 +72,8 @@ add_library(CommonFiles STATIC
     ui/MI_Text.cpp
     ui/MI_ScoreText.h
     ui/MI_ScoreText.cpp
+    ui/MI_Image.h
+    ui/MI_Image.cpp
 )
 
 if (USE_PNG_SAVE AND NOT USE_SDL2_LIBS)
