+++ src/common/ui/MI_ImageSelectField.cpp
@@ -1,6 +1,8 @@
 #include "MI_ImageSelectField.h"
 
 #include "ResourceManager.h"
+#include "ui/MI_Image.h"
+
 extern CResourceManager* rm;
 
 /**************************************
