+++ src/common/uicontrol.h
@@ -116,89 +116,4 @@ protected:
     short iControllingTeam = -1;
 };
 
-/*********************************************************
- *    Non Selectable Controls (Text and Images)
- *********************************************************/
-class MI_Image : public UI_Control
-{
-public:
-    MI_Image(gfxSprite * nspr, short x, short y, short srcx, short srcy, short w, short h, short numxframes, short numyframes, short speed);
-    virtual ~MI_Image();
-
-    void Update();
-    void Draw();
-
-    void SetPosition(short x, short y) {
-        ix = x;
-        iy = y;
-    }
-    void SetAnimationSpeed(short speed) {
-        iSpeed = speed;
-    }
-    void SetImage(short srcx, short srcy, short w, short h) {
-        isrcx = srcx;
-        isrcy = srcy;
-        iw = w;
-        ih = h;
-        iXFrame = srcx;
-        iYFrame = srcy;
-    }
-    void SetImageSource(gfxSprite * nspr) {
-        spr = nspr;
-    }
-
-    void SetPulse(bool pulse) {
-        fPulse = pulse;
-    }
-    void SetSwirl(bool swirl, float radius, float angle, float radiusSpeed, float angleSpeed) {
-        fSwirl = swirl;
-        dSwirlRadius = radius;
-        dSwirlAngle = angle;
-        dSwirlRadiusSpeed = radiusSpeed;
-        dSwirlAngleSpeed = angleSpeed;
-    }
-    void StopSwirl() {
-        SetSwirl(false, 0.0f, 0.0f, 0.0f, 0.0f);
-    }
-    void SetBlink(bool blink, short interval) {
-        fBlink = blink;
-        iBlinkInterval = interval;
-    }
-
-    bool IsSwirling() {
-        return fSwirl;
-    }
-
-    void GetPositionAndSize(short * x, short * y, short * w, short * h) {
-        *x = ix;
-        *y = iy;
-        *w = iw;
-        *h = ih;
-    }
-
-private:
-    gfxSprite * spr;
-
-    short iNumXFrames, iNumYFrames;
-    short isrcx, isrcy;
-    short iw, ih;
-    short iSpeed, iTimer;
-    short iXFrame, iYFrame;
-
-    bool fPulse;
-    short iPulseValue, iPulseDelay;
-    bool fPulseOut;
-
-    bool fSwirl;
-    float dSwirlRadius;
-    float dSwirlAngle;
-    float dSwirlRadiusSpeed;
-    float dSwirlAngleSpeed;
-
-    bool fBlink;
-    short iBlinkInterval;
-    short iBlinkCounter;
-    bool fBlinkShow;
-};
-
 #endif // UICONTROL_H
