+++ src/smw/menu/GameSettingsMenu.cpp
@@ -6,6 +6,7 @@
 #include "path.h"
 #include "ResourceManager.h"
 #include "ui/MI_Button.h"
+#include "ui/MI_Image.h"
 #include "ui/MI_Text.h"
 
 extern CGameMode * gamemodes[GAMEMODE_LAST];
