+++ src/smw/menu/MainMenu.cpp
@@ -2,6 +2,7 @@
 
 #include "ResourceManager.h"
 #include "ui/MI_Button.h"
+#include "ui/MI_Image.h"
 
 extern CResourceManager* rm;
