+++ src/smw/menu/MatchSelectionMenu.cpp
@@ -4,6 +4,7 @@
 #include "GameValues.h"
 #include "path.h"
 #include "ResourceManager.h"
+#include "ui/MI_Image.h"
 #include "ui/MI_Text.h"
 
 extern CGameValues game_values;
