+++ src/smw/menu/ModeOptionsMenu.cpp
@@ -4,6 +4,7 @@
 #include "GameMode.h"
 #include "ResourceManager.h"
 #include "ui/MI_Button.h"
+#include "ui/MI_Image.h"
 #include "ui/MI_Text.h"
 
 extern CGameValues game_values;
