+++ src/smw/menu/OptionsMenu.cpp
@@ -3,6 +3,7 @@
 #include "GameValues.h"
 #include "ResourceManager.h"
 #include "ui/MI_Button.h"
+#include "ui/MI_Image.h"
 #include "ui/MI_Text.h"
 
 extern CResourceManager* rm;
