+++ src/smw/menu/PlayerControlsSelectMenu.cpp
@@ -2,6 +2,7 @@
 
 #include "ResourceManager.h"
 #include "ui/MI_Button.h"
+#include "ui/MI_Image.h"
 #include "ui/MI_Text.h"
 
 extern CResourceManager* rm;
