+++ src/smw/menu/TourStopMenu.cpp
@@ -1,6 +1,7 @@
 #include "TourStopMenu.h"
 
 #include "ResourceManager.h"
+#include "ui/MI_Image.h"
 #include "ui/MI_Text.h"
 
 extern CResourceManager* rm;
