+++ src/smw/menu/TournamentScoreboardMenu.cpp
@@ -1,6 +1,7 @@
 #include "TournamentScoreboardMenu.h"
 
 #include "ResourceManager.h"
+#include "ui/MI_Image.h"
 
 extern CResourceManager* rm;
