+++ src/smw/menu/network/NetEditServersMenu.cpp
@@ -3,6 +3,8 @@
 #include "net.h"
 #include "ResourceManager.h"
 #include "ui/MI_Button.h"
+#include "ui/MI_Image.h"
+#include "ui/MI_StringScroll.h"
 #include "ui/MI_Text.h"
 #include "ui/MI_TextField.h"
