+++ src/smw/menu/network/NetLobbyMenu.cpp
@@ -2,6 +2,7 @@
 
 #include "net.h"
 #include "ResourceManager.h"
+#include "ui/MI_Image.h"
 #include "ui/MI_Text.h"
 #include "ui/MI_TextField.h"
