+++ src/smw/menu/network/NetNewRoomSettingsMenu.cpp
@@ -3,6 +3,7 @@
 #include "GameMode.h"
 #include "net.h"
 #include "ResourceManager.h"
+#include "ui/MI_Image.h"
 #include "ui/MI_Text.h"
 
 extern CResourceManager* rm;
