+++ src/smw/menu/network/NetRoomMenu.cpp
@@ -4,6 +4,7 @@
 #include "net.h"
 #include "path.h"
 #include "ResourceManager.h"
+#include "ui/MI_Image.h"
 #include "ui/MI_Text.h"
 #include "ui/MI_TextField.h"
