+++ src/smw/menu/options/GraphicsOptionsMenu.cpp
@@ -3,6 +3,7 @@
 #include "FileList.h"
 #include "GameValues.h"
 #include "ResourceManager.h"
+#include "ui/MI_Image.h"
 #include "ui/MI_Text.h"
 
 extern CResourceManager* rm;
