+++ src/smw/menu/options/PowerupDropRatesMenu.cpp
@@ -1,6 +1,7 @@
 #include "PowerupDropRatesMenu.h"
 
 #include "ResourceManager.h"
+#include "ui/MI_Image.h"
 #include "ui/MI_Text.h"
 
 extern CResourceManager* rm;
