+++ src/smw/menu/options/ProjectileLimitsMenu.cpp
@@ -2,6 +2,7 @@
 
 #include "uicontrol.h"
 #include "ui/MI_Button.h"
+#include "ui/MI_Image.h"
 #include "ui/MI_SelectField.h"
 #include "ui/MI_Text.h"
