+++ src/smw/ui/MI_FrenzyModeOptions.cpp
@@ -3,6 +3,7 @@
 #include "GameValues.h"
 #include "ResourceManager.h"
 #include "ui/MI_Button.h"
+#include "ui/MI_Image.h"
 
 extern CGameValues game_values;
 extern CResourceManager* rm;
