+++ src/smw/ui/MI_InputControlContainer.cpp
@@ -3,6 +3,7 @@
 #include "GameValues.h"
 #include "gfx/gfxSprite.h"
 #include "ui/MI_Button.h"
+#include "ui/MI_Image.h"
 #include "ui/MI_SelectField.h"
 #include "ui/MI_Text.h"
 #include "ResourceManager.h"
