+++ src/smw/ui/MI_TeamSelect.cpp
@@ -3,6 +3,7 @@
 #include "GameValues.h"
 #include "ResourceManager.h"
 #include "FileList.h"
+#include "ui/MI_Image.h"
 
 extern CGameValues game_values;
 extern CResourceManager* rm;
